backend/requirements.txt
Django==4.1.3
djangorestframework
django-cors-headers
load_dotenv
dj-config-url
psycopg2